examples/image/requirements.txt
submitit
torchmetrics[image]
torchvision
wandb
